Move scripts out of src
