+++ build-scripts/generate-epubs.ts
@@ -7,12 +7,12 @@ import remarkTwoslash from "remark-shiki-twoslash";
 import rehypeRaw from "rehype-raw";
 import rehypeSlug from "rehype-slug-custom-id";
 import { UserConfigSettings } from "shiki-twoslash";
-import { collections, unicorns } from "utils/data";
-import { getAllPosts } from "utils/get-all-posts";
+import { collections, unicorns } from "../src/utils/data";
+import { getAllPosts } from "../src/utils/get-all-posts";
 import { join, resolve } from "path";
 import { visit } from "unist-util-visit";
 import { Element, Root } from "hast";
-import { isRelativePath } from "utils/url-paths";
+import { isRelativePath } from "../src/utils/url-paths";
 import { EPub } from "@lesjoursfr/html-to-epub";
 import { unified } from "unified";
 import { CollectionInfo, PostInfo, RawCollectionInfo } from "types/index";
