+++ build-scripts/social-previews/read-file-as-base64.ts
