+++ package.json
@@ -29,10 +29,10 @@
 		"host:local": "cd dist && ws --http2 --compress",
 		"format": "prettier -w . --cache --plugin-search-dir=.",
 		"lint": "eslint . --ext .js,.ts,.astro",
-		"search-index": "tsx src/build-scripts/search-index.ts",
-		"social-previews:build": "node node_modules/puppeteer-core/install.js && tsx --tsconfig tsconfig.script.json src/build-scripts/social-previews/index.ts",
-		"social-previews:dev": "tsx --tsconfig tsconfig.script.json src/build-scripts/social-previews/live-server.ts",
-		"epub": "tsx --tsconfig tsconfig.script.json src/build-scripts/generate-epubs.ts",
+		"search-index": "tsx build-scripts/search-index.ts",
+		"social-previews:build": "node node_modules/puppeteer-core/install.js && tsx --tsconfig tsconfig.script.json build-scripts/social-previews/index.ts",
+		"social-previews:dev": "tsx --tsconfig tsconfig.script.json build-scripts/social-previews/live-server.ts",
+		"epub": "tsx --tsconfig tsconfig.script.json build-scripts/generate-epubs.ts",
 		"tsc": "tsc --noEmit && astro check",
 		"test": "jest",
 		"prepare": "husky install"
